This project is based on Neural Style Transfer of one image(style image) to another image(content image).  Here, we have transferred the style and art of Monalisa image to Rabindranath Tagore image using transfer learning  (with pre-trained (trained on imagenet dataset) convolutional neural network (VGG19) of Google). content_weight was taken as 1 and style_weight was taken as 10. After 500 epochs result was very encouraging and the style was successfully transferred from style image (Monalisa) to the content image (Rabindranath Tagore) and the new image of Rabindranath Tagore was very artistic with the art of Monalisa image.
This project is based on Neural Style Transfer of one image(style image) to another image(content image).
Here, we have transferred the style and art of Monalisa image to Rabindranath Tagore image using transfer learning
(with pre-trained (trained on imagenet dataset) convolutional neural network (VGG19) of Google). content_weight was taken as 1 and style_weight was taken as 10.
After 500 epochs result was very encouraging and the style was successfully transferred from style image (Monalisa) to the content image (Rabindranath Tagore) and the new image of Rabindranath Tagore was very artistic with the art of Monalisa image.